+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "api-spec-template"]
-	path = api-spec-template
-	url = git@github.com:Metatavu/api-spec-template.git
+[submodule "rapurc-api-spec"]
+	path = rapurc-api-spec
+	url = git@github.com:Metatavu/rapurc-api-spec.git
