+++ src/main/resources/db/changeLog.xml
@@ -3,30 +3,5 @@
                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
 
-    <changeSet id="examples" author="jari.nykanen">
-        <createTable tableName="example">
-            <column name="id" type="binary(16)">
-                <constraints nullable="false" primaryKey="true"/>
-            </column>
-            <column name="name" type="varchar(191)">
-                <constraints nullable="false"/>
-            </column>
-            <column name="amount" type="int">
-                <constraints nullable="false"/>
-            </column>
-            <column name="createdat" type="datetime(6)">
-                <constraints nullable="false"/>
-            </column>
-            <column name="modifiedat" type="datetime(6)">
-                <constraints nullable="false"/>
-            </column>
-            <column name="creatorid" type="binary(16)">
-                <constraints nullable="false"/>
-            </column>
-            <column name="lastmodifierid" type="binary(16)">
-                <constraints nullable="false"/>
-            </column>
-        </createTable>
-    </changeSet>
 
 </databaseChangeLog>
