--- src/test/kotlin/fi/metatavu/example/api/test/common/moshi/adapters/OffsetDateTimeAdapter.kt
@@ -1,24 +1,0 @@
-package fi.metatavu.example.api.test.common.moshi.adapters
-
-import com.squareup.moshi.FromJson
-import com.squareup.moshi.ToJson
-import java.time.OffsetDateTime
-import java.time.format.DateTimeFormatter
-
-/**
- * Adapter class for OffsetDateTime
- *
- * @author Jari Nykänen
- */
-class OffsetDateTimeAdapter {
-
-    @ToJson
-    fun toJson(value: OffsetDateTime): String {
-        return DateTimeFormatter.ISO_OFFSET_DATE_TIME.format(value)
-    }
-
-    @FromJson
-    fun fromJson(value: String): OffsetDateTime {
-        return OffsetDateTime.parse(value, DateTimeFormatter.ISO_OFFSET_DATE_TIME)
-    }
-}
