--- src/test/kotlin/fi/metatavu/example/api/test/functional/resources/LocalTestProfile.kt
@@ -1,7 +1,0 @@
-package fi.metatavu.example.api.test.functional.resources
-
-import io.quarkus.test.junit.QuarkusTestProfile
-
-class LocalTestProfile: QuarkusTestProfile {
-
-}
