--- src/test/kotlin/fi/metatavu/example/api/test/functional/settings/ApiTestSettings.kt
@@ -1,16 +1,0 @@
-package fi.metatavu.example.api.test.functional.settings
-
-/**
- * Settings implementation for test builder
- *
- * @author Jari Nykänen
- */
-class ApiTestSettings {
-
-    /**
-     * Returns API service base path
-     */
-    val apiBasePath: String
-        get() = "http://localhost:8081"
-
-}
