+++ src/test/kotlin/fi/metatavu/rapurc/api/test/functional/resources/LocalTestProfile.kt
@@ -1,0 +1,7 @@
+package fi.metatavu.rapurc.api.test.functional.resources
+
+import io.quarkus.test.junit.QuarkusTestProfile
+
+class LocalTestProfile: QuarkusTestProfile {
+
+}
