+++ src/test/resources/kc.json
@@ -1,6 +1,6 @@
 {
-  "id" : "Edufication",
-  "realm" : "Edufication",
+  "id" : "rapurc",
+  "realm" : "rapurc",
   "notBefore" : 0,
   "revokeRefreshToken" : false,
   "refreshTokenMaxReuse" : 0,
@@ -47,14 +47,14 @@
       "description" : "${role_offline-access}",
       "composite" : false,
       "clientRole" : false,
-      "containerId" : "Edufication",
+      "containerId" : "rapurc",
       "attributes" : { }
     }, {
       "id" : "62af1558-7dd2-4ae8-ad00-2a0b22f7bcd3",
       "name" : "manager",
       "composite" : false,
       "clientRole" : false,
-      "containerId" : "Edufication",
+      "containerId" : "rapurc",
       "attributes" : { }
     }, {
       "id" : "cd378218-f2c3-4532-bf08-75af3a24dd05",
@@ -62,7 +62,7 @@
       "description" : "${role_uma_authorization}",
       "composite" : false,
       "clientRole" : false,
-      "containerId" : "Edufication",
+      "containerId" : "rapurc",
       "attributes" : { }
     } ],
     "client" : {
@@ -388,14 +388,14 @@
     "clientId" : "account",
     "name" : "${client_account}",
     "rootUrl" : "${authBaseUrl}",
-    "baseUrl" : "/realms/Edufication/account/",
+    "baseUrl" : "/realms/rapurc/account/",
     "surrogateAuthRequired" : false,
     "enabled" : true,
     "alwaysDisplayInConsole" : false,
     "clientAuthenticatorType" : "client-secret",
     "secret" : "e275b08c-8169-4716-8009-4dfdf02cd0aa",
     "defaultRoles" : [ "manage-account", "view-profile" ],
-    "redirectUris" : [ "/realms/Edufication/account/*" ],
+    "redirectUris" : [ "/realms/rapurc/account/*" ],
     "webOrigins" : [ ],
     "notBefore" : 0,
     "bearerOnly" : false,
@@ -418,13 +418,13 @@
     "clientId" : "account-console",
     "name" : "${client_account-console}",
     "rootUrl" : "${authBaseUrl}",
-    "baseUrl" : "/realms/Edufication/account/",
+    "baseUrl" : "/realms/rapurc/account/",
     "surrogateAuthRequired" : false,
     "enabled" : true,
     "alwaysDisplayInConsole" : false,
     "clientAuthenticatorType" : "client-secret",
     "secret" : "d8152e0e-5a8e-4f0c-9a01-c838ed206570",
-    "redirectUris" : [ "/realms/Edufication/account/*" ],
+    "redirectUris" : [ "/realms/rapurc/account/*" ],
     "webOrigins" : [ ],
     "notBefore" : 0,
     "bearerOnly" : false,
@@ -581,13 +581,13 @@
     "clientId" : "security-admin-console",
     "name" : "${client_security-admin-console}",
     "rootUrl" : "${authAdminUrl}",
-    "baseUrl" : "/admin/Edufication/console/",
+    "baseUrl" : "/admin/rapurc/console/",
     "surrogateAuthRequired" : false,
     "enabled" : true,
     "alwaysDisplayInConsole" : false,
     "clientAuthenticatorType" : "client-secret",
     "secret" : "498c10ee-6d0a-497a-b619-369b132997ac",
-    "redirectUris" : [ "/admin/Edufication/console/*" ],
+    "redirectUris" : [ "/admin/rapurc/console/*" ],
     "webOrigins" : [ "+" ],
     "notBefore" : 0,
     "bearerOnly" : false,
